fix process launch and kill for windows.

+++ pythymiodw/pyro/__init__.py
@@ -1,4 +1,5 @@
 import os
+import sys
 import Pyro4
 import subprocess
 import signal
@@ -9,13 +10,21 @@ from pythymiodw.io import ProxGround
 
 class ThymioMR():
 	def __init__(self):
-		self.pyro4daemon_proc=subprocess.Popen(['python -m pythymiodw.pyro.__main__'], stdout=subprocess.PIPE, shell=True, preexec_fn=os.setsid)    
+		if sys.platform == 'win32':
+			self.pyro4daemon_proc=subprocess.Popen([sys.executable,'-m', 'pythymiodw.pyro.__main__'], stdout=subprocess.PIPE, shell=True)
+		else:
+			self.pyro4daemon_proc=subprocess.Popen(['python -m pythymiodw.pyro.__main__'], stdout=subprocess.PIPE, shell=True, preexec_fn=os.setsid)    
+
 		time.sleep(2)
 		self.robot = Pyro4.Proxy('PYRONAME:pythymiodw.thymiosimmr')
 
 	def quit(self):
 		self.robot.quit()
-		os.killpg(os.getpgid(self.pyro4daemon_proc.pid), signal.SIGTERM)        
+		if sys.platform == 'win32':
+			subprocess.call(['taskkill','/F','/T','/PID',str(self.pyro4daemon_proc.pid)])
+		else:
+			os.killpg(os.getpgid(self.pyro4daemon_proc.pid), signal.SIGTERM)        
+
 
 	def wheels(self, lv, rv):
 		self.robot.wheels(lv, rv)
